Prohibit change of JValue returned from operator[]
:Release Notes:
Prohibit change of JValue object that is returned from JValue::operator[]

:Detailed Notes:
The json library has other interfaces to change the internal structure or values of JValue object.
Due to the fact that JValue is just a reference to a real json object, using of the bracket
operator can confuse users of the library.

:Testing Performed:
Starfish image was built to detect if other projects were not affected by the change.
It was tested using internal unit tests.

:QA Notes:

:Issues Addressed:
[GF-6380] libpbnjson: Assigning to array or object with bracket operator is not working

Open-webOS-DCO-1.0-Signed-off-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>

Change-Id: I932e30d1894d7e3ce20231fc36590709d5df01cc
Reviewed-on: https://g2g.palm.com/2809
Reviewed-by: Build Verification
Reviewed-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>
Tested-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>
Reviewed-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>