Prevent modification of schema passed to resolver
:Release Notes:
Prevent modification of a schema object, that is passed to the resolver

:Detailed Notes:
Schema that is passed to reslover was modified, as a result it was not
possible to reuse the same schema object.

:Testing Performed:
One more unit test was written.
All internal unit tests were passed successfully.
Starfish-image was built and checked on virtualbox

:QA Notes:

:Issues Addressed:
[GF-7879] pbnjson resolve fail if pbnjson::JSchema is reused

Open-webOS-DCO-1.0-Signed-off-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>

Change-Id: Ie76969732fecaf2d799805b7e401250bb8d38687
Reviewed-on: https://g2g.palm.com/2839
Reviewed-by: Build Verification
Reviewed-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>
Tested-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>